Refactor the SimDB exporters (#572)
* Refactor the SimDB exporters

* Refactor the SimDB exporters

* Remove dead code

* Touch script

* Refactor the SimDB exporters

* Refactor the SimDB exporters

* Refactor the SimDB exporters
