+++ sparta/scripts/simdb/exporters/export_text_report.py
@@ -1,83 +1,7 @@
-import io, os, zlib, struct
+from .sim_utils import *
+from .stat_utils import *
+import io
 from .utils import FormatNumber
-
-class SimulationInfo:
-    @classmethod
-    def GetHeaderPairs(cls, db_conn):
-        cursor = db_conn.cursor()
-        cmd = "SELECT HeaderName, HeaderValue FROM SimulationInfoHeaderPairs"
-        cursor.execute(cmd)
-
-        header_pairs = []
-        for p in cursor.fetchall():
-            header_pairs.append((p[0], p[1]))
-
-        return header_pairs
-
-    @classmethod
-    def Stringize(cls, db_conn, out, line_start = "# ", line_end = "\n", show_field_names = True):
-        pairs = cls.GetHeaderPairs(db_conn)
-        for p in pairs:
-            out.write(line_start)
-            if show_field_names:
-                # out << std::left << std::setw(10) << (p.first + ":");
-                name = p[0] + ":"
-                out.write(f"{name:<10}")
-
-            # out << p.second << line_end;
-            out.write(f"{p[1]}{line_end}")
-
-class StatisticInstance:
-    def __init__(self, name, location, desc, stat_value_getter):
-        self.name = name
-        self.location = location
-        self.desc = desc
-        self.stat_value_getter = stat_value_getter
-
-    def GetName(self):
-        return self.name
-    
-    def GetLocation(self):
-        return self.location
-
-    def GetDesc(self, unused=False):
-        assert unused == False
-        return self.desc
-
-    def GetValue(self):
-        return self.stat_value_getter.GetNext()
-
-class StatValueGetter:
-    def __init__(self, stats_values):
-        self.stats_values = stats_values
-        self.index = 0
-
-    def GetNext(self):
-        if self.index >= len(self.stats_values):
-            raise IndexError("No more values in stats blob")
-        value = self.stats_values[self.index]
-        self.index += 1
-        return value
-
-def GetStatsValuesGetter(cursor, dest_file):
-    dest_file_name = os.path.basename(dest_file)
-    cmd = f"SELECT Data, IsCompressed FROM CollectionRecords WHERE Notes='{dest_file_name}'"
-    cursor.execute(cmd)
-    stats_blob, is_compressed = cursor.fetchone()
-    if is_compressed:
-        stats_blob = zlib.decompress(stats_blob)
-
-    # Turn the stats blob (byte vector) into a vector of doubles.
-    assert len(stats_blob) % (2+8) == 0, "Invalid stats blob length"
-    stats_values = []
-    for i in range(0, len(stats_blob), 2+8):
-        # The first value is the collectable ID, the second is the value.
-        # We don't need the collectable ID here.
-        val = struct.unpack("d", stats_blob[i+2:i+10])[0]
-        val = FormatNumber(val, as_string=False)
-        stats_values.append(val)
-
-    return StatValueGetter(stats_values)
 
 class TextReportExporter:
     DEFAULT_REPORT_PREFIX = "Report "
@@ -158,15 +82,8 @@ class TextReportExporter:
 
         #const bool show_descs =
         #    (r->getStyle("show_descriptions", getShowDescriptions() ? "true" : "false") == "true");
-        cmd = f"SELECT StyleValue FROM ReportStyles WHERE ReportDescID = {descriptor_id} AND ReportID = {report_id} AND StyleName = 'show_descriptions'"
-        cursor.execute(cmd)
-
-        show_descs = cursor.fetchone()
-        if show_descs is None:
-            show_descs = self.GetShowDescriptions()
-        else:
-            assert show_descs[0].lower() in ("true", "false")
-            show_descs = show_descs[0].lower() == "true"
+        default_show_descs = "true" if self.GetShowDescriptions() else "false"
+        show_descs = GetReportStyle(cursor, report_id, descriptor_id, "show_descriptions", default_show_descs) == "true"
 
         # const std::string INDENT_STR = "  ";
         indent_str = "  "
@@ -175,7 +92,7 @@ class TextReportExporter:
         additional_stat_indent = "  "
 
         # if(write_contentless_reports_ || hasStatistics_(r)){
-        if self.GetWriteContentlessReports() or self.__HasStatistics(db_conn, report_id):
+        if self.GetWriteContentlessReports() or HasStatistics(db_conn.cursor(), report_id):
             # std::stringstream indent;
             # if(indent_subreports_){
             #     for(uint32_t d=0; d<depth; ++d){
@@ -237,7 +154,7 @@ class TextReportExporter:
                 val_col_after_indent -= len(indent)
 
             # for(const statistics::stat_pair_t& si : r->getStatistics()){
-            report_stats = self.__GetReportStatistics(db_conn, report_id)
+            report_stats = GetReportStatInsts(db_conn.cursor(), report_id, self.stat_value_getter)
             for si in report_stats:
                 si_name = si.GetName()
                 if not si_name:
@@ -273,7 +190,7 @@ class TextReportExporter:
 
                 # // Print description column
                 # TODO cnyce: Fill in this code!
-                assert self.GetShowDescriptions() == False, "Description column not implemented yet"
+                assert show_descs == False, "Description column not implemented yet"
 
                 # // Write line to the output
                 # out << name.str() << std::endl;
@@ -290,7 +207,7 @@ class TextReportExporter:
             # for(const Report& sr : r->getSubreports()){
             #     dump_(out, &sr, depth+1);
             # }
-            subrep_ids = self.__GetSubreports(db_conn, descriptor_id, report_id)
+            subrep_ids = GetSubreportIDs(db_conn.cursor(), descriptor_id, report_id)
             for subrep_id in subrep_ids:
                 self.__Dump(db_conn, out, descriptor_id, subrep_id, depth + 1)
 
@@ -301,32 +218,3 @@ class TextReportExporter:
             # }
             if len(subrep_ids) > 0:
                 out.write("\n")
-
-    def __HasStatistics(self, db_conn, report_id):
-        cursor = db_conn.cursor()
-        cmd = "SELECT COUNT(*) FROM StatisticInsts WHERE ReportID = ?"
-        cursor.execute(cmd, (report_id,))
-        count = cursor.fetchone()[0]
-        return count > 0
-
-    def __GetReportStatistics(self, db_conn, report_id):
-        cursor = db_conn.cursor()
-        cmd = "SELECT StatisticName, StatisticLoc, StatisticDesc FROM StatisticInsts WHERE ReportID = ?"
-        cursor.execute(cmd, (report_id,))
-
-        statistics = []
-        for name, loc, desc in cursor.fetchall():
-            statistics.append(StatisticInstance(name, loc, desc, self.stat_value_getter))
-
-        return statistics
-
-    def __GetSubreports(self, db_conn, descriptor_id, parent_report_id):
-        cmd = f"SELECT Id FROM Reports WHERE ReportDescID = {descriptor_id} AND ParentReportID = {parent_report_id}"
-        cursor = db_conn.cursor()
-        cursor.execute(cmd)
-
-        subrep_ids = []
-        for row in cursor.fetchall():
-            subrep_ids.append(row[0])
-
-        return subrep_ids
